Basic FrameWork ready and tested
1) lots of enhancements and usability bugs fixed.
2) suite/individual test case can be replayed.
3) separate logs for relay and monitor.
4) writing to test_case.json file sequencing issues fixed

+++ lauch_file.py
@@ -1,5 +1,7 @@
 import os
 from topo import dut_connections as dcon
+from topo import g_curr_tc_name_holder
+
 import click
 
 """
@@ -14,10 +16,24 @@ And execute the appropriate command on the terminal.
 @click.group('launchmr')
 @click.pass_context
 @click.option('-g', '--gdb', is_flag=True, help='stop at pdb on error')
-def launch_mr(ctx, gdb):
+@click.option('--test_suite', help='test suite name')
+@click.option('--tc_name', help='test suite name')
+def launch_mr(ctx, gdb, test_suite='', tc_name=''):
     ctx.obj['options'] = ''
+    ctx.obj['tc_name'] = ''
+    ctx.obj['test_suite'] = ''
+
     if gdb:
         ctx.obj['options'] += ' --gdb'
+
+    # ctx.obj['test_suite'] = 'watch_lists/' + input('Test Suite Name : watch_lists/')
+    if tc_name:
+        ctx.obj['tc_name'] = tc_name
+    elif test_suite:
+        ctx.obj['test_suite'] = test_suite
+    else:
+        print('No test suite or tc_name given ')
+        exit(-1)
     pass
 
 
@@ -25,28 +41,54 @@ def launch_mr(ctx, gdb):
 @click.pass_context
 def launch_mr_monitor(ctx):
     # erase contents of the file
-    open('watch_list.json', 'w').close()
+
+    if not ctx.obj['test_suite']:
+        print('Monitor cant start. Test-suite name is not given')
+        exit(-1)
+
+    options = ctx.obj['options']
+    options += ' --test_suite ' + ctx.obj['test_suite']
+
+    if os.path.exists(ctx.obj['test_suite']):
+        if 'n' == input('Test Suite already exists, overwrite? [y/n]').strip():
+            exit(1)
+
+    with open(ctx.obj['test_suite'], 'w') as f:
+        fname = '{}_init.json'.format(ctx.obj['test_suite'])
+        f.writelines('\n')  # start with \n for simplifiying last line read
+        f.writelines(fname)
+    with open(g_curr_tc_name_holder, 'w') as f:
+        f.writelines(fname)
+
+    open('{}_init.json'.format(ctx.obj['test_suite']), 'w').close()
     open('logs/mrlog.log', 'w').close()
 
     for dut in dcon.keys():
         with open('{}_monitor.sh'.format(dut), 'w') as d1bash:
             d1bash.writelines('#!/bin/bash\
             \n\ncd /Users/dr412113/PycharmProjects/monitorandreplay\
-            \necho connect to {}\npython3.7 ./mr.py {} monitor {} {}\
-            \nbash'.format(dut, ctx.obj['options'], dcon[dut]['ip'], dcon[dut]['port']))
+            \necho connect to {}\npython3.7 ./mr.py {} monitor {}\
+            \nbash\n'.format(dut, options, dcon[dut]['ip']))
 
-        os.system('chmod +x %s_monitor.sh'%(dut))
-        os.system('open -a Terminal %s_monitor.sh'%(dut))
+        os.system('chmod +x %s_monitor.sh' % dut)
+        os.system('open -a Terminal %s_monitor.sh' % dut)
 
 
 @launch_mr.command('replay')
 @click.pass_context
 def launch_mr_replay(ctx):
+    options = ctx.obj['options']
+
+    if ctx.obj['tc_name']:
+        options += '--tc_name ' + ctx.obj['tc_name']
+    else:
+        options += '--test_suite ' + ctx.obj['test_suite']
+
     with open('replay.sh', 'w') as d1bash:
         d1bash.writelines('#!/bin/bash\
         \n\ncd /Users/dr412113/PycharmProjects/monitorandreplay\
         \npython3.7 ./mr.py {} replay\
-        \nbash'.format(ctx.obj['options']))
+        \nbash'.format(options))
     os.system('chmod +x replay.sh')
     os.system('open -a Terminal replay.sh')
 
@@ -54,4 +96,3 @@ def launch_mr_replay(ctx):
 if __name__ == '__main__':
     launch_mr(obj={})
     pass
-
